# specs/features/plugacc/api.yaml
...
openapi: 3.1.0
...
info:
  title: The administrator and the enablement of a plugin
  version: 0.1.0
servers:
  - url: /
    description: The hub.
security:
  - sessionCookie: []
paths:
...
  /workspaces/{workspaceId}/plugins:
    parameters:
      - $ref: "../../api/components.yaml#/components/parameters/WorkspaceId"
    get:
      operationId: listEnabledPlugins
      summary: List the plugins that the workspace enables.
      description: >
        Realizes `PLUGACC-FR-017`. The permission is `workspace.read`. Admits an
        administrator. One page holds every item, ordered by the plugin
        identifier. `DAT-009`. Each item carries the entry of the plugin, whatever
        the allowlist of the member holds. `PLUGACC-DD-009`.
      parameters:
        - $ref: "../../api/components.yaml#/components/parameters/FlowId"
      responses:
        "200":
          headers:
            X-Flow-ID:
              $ref: "../../api/components.yaml#/components/headers/FlowId"
          description: One page of the enablements.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "../../api/components.yaml#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/EnabledPlugin"
        "401":
          $ref: "../../api/components.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "../../api/components.yaml#/components/responses/Forbidden"
        "404":
          $ref: "../../api/components.yaml#/components/responses/NotFound"
  /workspaces/{workspaceId}/plugins/{pluginId}:
    parameters:
      - $ref: "../../api/components.yaml#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/PluginId"
    get:
      operationId: getEnabledPlugin
      summary: Read one enablement.
      description: >
        Realizes `PLUGACC-FR-017`. The permission is `workspace.read`. A plugin that
        the workspace does not enable answers `not-found`.
      parameters:
        - $ref: "../../api/components.yaml#/components/parameters/FlowId"
      responses:
        "200":
          headers:
            X-Flow-ID:
              $ref: "../../api/components.yaml#/components/headers/FlowId"
          description: The enablement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnabledPlugin"
        "401":
          $ref: "../../api/components.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "../../api/components.yaml#/components/responses/Forbidden"
        "404":
          $ref: "../../api/components.yaml#/components/responses/NotFound"
    put:
      operationId: enablePlugin
      summary: Enable a plugin in the workspace.
      description: >
        Realizes `PLUGACC-FR-012` and `PLUGACC-FR-013`. The permission is
        `plugins.write`. A manager enables a plugin of their allowlist, and an
        administrator any loaded plugin. A plugin off the allowlist answers
        `/problems/http/permission-denied` with the permission `plugin-allowlist`.
      parameters:
        - $ref: "../../api/components.yaml#/components/parameters/FlowId"
      responses:
        "200":
          headers:
            X-Flow-ID:
              $ref: "../../api/components.yaml#/components/headers/FlowId"
          description: The workspace already enabled the plugin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnabledPlugin"
        "201":
          headers:
            X-Flow-ID:
              $ref: "../../api/components.yaml#/components/headers/FlowId"
          description: The workspace enables the plugin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnabledPlugin"
        "401":
          $ref: "../../api/components.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "../../api/components.yaml#/components/responses/Forbidden"
        "404":
          $ref: "../../api/components.yaml#/components/responses/NotFound"
        "422":
          $ref: "../../api/components.yaml#/components/responses/Invalid"
    delete:
      operationId: disablePlugin
      summary: Disable a plugin in the workspace. Its records and settings stay.
      description: >
        Realizes `PLUGACC-FR-014`, `PLUGACC-FR-015`, and `PLUGACC-FR-027`. The
        permission is `plugins.write`.
      parameters:
        - $ref: "../../api/components.yaml#/components/parameters/FlowId"
      responses:
        "204":
          headers:
            X-Flow-ID:
              $ref: "../../api/components.yaml#/components/headers/FlowId"
          description: The workspace no longer enables the plugin.
        "401":
          $ref: "../../api/components.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "../../api/components.yaml#/components/responses/Forbidden"
        "404":
          $ref: "../../api/components.yaml#/components/responses/NotFound"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-maroid_session
      description: >
        The access token that Dex issued. `SEC-005` gives this cookie as the one
        path of a web request, and the hub reads no `Authorization` header on it.
  parameters:
...
    PluginId:
      name: pluginId
      in: path
      required: true
      description: The plugin identifier. See `PLG-003`.
      schema:
        type: string
        examples: [dev.maroid.jasmine]
  schemas:
...
    EnabledPlugin:
      type: object
      required: [plugin_id, created_at]
      description: >
        A plugin that the workspace enables. `plugin` is absent when the hub did not
        load the plugin at its start. `PLUGACC-DD-009`.
      properties:
        plugin_id:
          type: string
          description: The plugin identifier. See `PLG-003`.
        created_at:
          type: string
          format: date-time
          readOnly: true
        plugin:
          $ref: "../pcap/api.yaml#/components/schemas/PluginEntry"
...